refactor(tests): move cli-independent docker tests to test_scan_docker.py
